# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "kasrchem System",
        "description": "kasrchem API Documentation",
        "contact": {
            "email": "admin@kasrchem.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "https://backend.kasrchem.murabba.dev/api/v2",
            "description": "API Server"
        }
    ],
    "paths": {
...
        "/pages": {
            "get": {
                "tags": [
                    "pages"
                ],
                "summary": "Get list of Pages",
                "description": "Returns list of Pages",
                "operationId": "Get list of Pages",
                "parameters": [
                    {
                        "name": "locale",
                        "in": "header",
                        "description": "App Locale",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "default": "ar",
                            "enum": [
                                "ar",
                                "en"
                            ]
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "404": {
                        "description": "not found"
                    }
                }
            }
        },
        "/page/{id}": {
            "get": {
                "tags": [
                    "pages"
                ],
                "summary": "Get page data",
                "description": "Returns page data",
                "operationId": "getPagedata",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Page ID",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    },
                    {
                        "name": "locale",
                        "in": "header",
                        "description": "App Locale",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "default": "ar",
                            "enum": [
                                "ar",
                                "en"
                            ]
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "404": {
                        "description": "not found"
                    }
                }
            }
        },
...
        "/posts": {
            "get": {
                "tags": [
                    "Multimedia"
                ],
                "summary": "Get list of posts",
                "operationId": "getPostsList",
                "parameters": [
                    {
                        "name": "locale",
                        "in": "header",
                        "description": "App Locale",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "default": "ar",
                            "enum": [
                                "ar",
                                "en"
                            ]
                        }
                    },
                    {
                        "name": "category_id",
                        "in": "query",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "404": {
                        "description": "not found"
                    }
                }
            }
        },
        "/posts/{id}": {
            "get": {
                "tags": [
                    "Multimedia"
                ],
                "summary": "Show Post Data",
                "operationId": "show_post",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Post ID",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    },
                    {
                        "name": "locale",
                        "in": "header",
                        "description": "App Locale",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "default": "ar",
                            "enum": [
                                "ar",
                                "en"
                            ]
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "404": {
                        "description": "not found"
                    }
                }
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "Bearer": {
                "type": "apiKey",
                "description": "Pass Bearer Token",
                "name": "Authorization",
                "in": "header"
            }
        }
    }
}